requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-heroku==0.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
razorpay==1.4.2
requests==2.32.3
setuptools==70.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
